+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.12.4 (unreleased)
 --------------------
-* update OpenSSL to 1.0.2h
+* add support for building with VS2015
+* update OpenSSL to 1.0.2j and libpng to 1.6.26
 * **[qt]** change screen size in headless mode to a more sane value (1366x768)
 * **[qt]** allow screen size in headless mode to be overriden via environment variables
 * **[qt]** fix compilation failure on OS X 10.11 due to usage of deprecated APIs